feat: a few tests for compaction
